+++ models/listing.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 const Review = require("./review.js");
+const User = require("./user.js");
 
 const listingSchema = new Schema({
     title: {
@@ -39,6 +40,10 @@ const listingSchema = new Schema({
             ref: "Review",
         },
     ],
+    owner: {
+        type: Schema.Types.ObjectId,
+        ref: "User"
+    }
 });
