+++ routes/listing.js
@@ -4,21 +4,21 @@ const router = express.Router();
 
 const Listing = require("../models/listing.js");
 const wrapAsync = require("../utils/wrapAsync.js");
-const ExpressError = require("../utils/ExpressError.js");
-const { listingSchema } = require("../schema.js");
-const { isLoggedIn } = require("../middleware.js");
+// const ExpressError = require("../utils/ExpressError.js");
+// const { listingSchema } = require("../schema.js");             // moved with middleware
+const { isLoggedIn, isOwner, validateListing } = require("../middleware.js");
 
 
-//middleware for validating listing data using Joi schema
-const validateListing = (req, res, next) => {
-    let { error } = listingSchema.validate(req.body);
-    if (error) {
-        let errMsg = error.details.map((el) => el.message).join(",");
-        throw new ExpressError(400, errMsg);
-    } else {
-        next();
-    }
-}
+// //middleware for validating listing data using Joi schema
+// const validateListing = (req, res, next) => {
+//     let { error } = listingSchema.validate(req.body);
+//     if (error) {
+//         let errMsg = error.details.map((el) => el.message).join(",");
+//         throw new ExpressError(400, errMsg);
+//     } else {
+//         next();
+//     }
+// }                       // moved to dedicated middleware.js file and imported
 
 
 //Index Route
@@ -40,11 +40,12 @@ router.get("/new", isLoggedIn, (req, res) => {                            // now
 //Show Route
 router.get("/:id", wrapAsync(async (req, res) => {
     let { id } = req.params;
-    let listing = await Listing.findById(id).populate("reviews");                                      // previously(before reviews) : let listing = await Listing.findById(id);
+    let listing = await Listing.findById(id).populate("reviews").populate("owner");                                      // previously(before reviews) : let listing = await Listing.findById(id);            // also chain the owner field to be populated
     if (!listing) {
         req.flash("error", "Listing does not exist!");
-        return res.redirect("/listings");                                   
+        return res.redirect("/listings");
     }
+    // console.log(listing);
     // // alt:  throwing an error and directing to the error page with a status code
     // if (!listing) {
     //     throw new ExpressError(404, "Listing does not exist!");
@@ -82,13 +83,16 @@ router.post("/", validateListing, wrapAsync(async (req, res, next) => {
     // }
     //     now converting the above joi related code into a middleware by converting it into a function
     const newListing = new Listing(req.body.listing);
+    //
+    newListing.owner = req.user._id;
+    //
     await newListing.save();
     req.flash("success", "New Listing Created!");                // before redirecting,... success local variable is then used in index page of listings
     res.redirect("/listings");
 }));
 
 //Edit Route
-router.get("/:id/edit", isLoggedIn, wrapAsync(async (req, res) => {
+router.get("/:id/edit", isLoggedIn, isOwner, wrapAsync(async (req, res) => {
     let { id } = req.params;
     let oldListing = await Listing.findById(id);
     if (!oldListing) {
@@ -99,18 +103,25 @@ router.get("/:id/edit", isLoggedIn, wrapAsync(async (req, res) => {
 }));
 
 //Update Route
-router.put("/:id", isLoggedIn, validateListing, wrapAsync(async (req, res) => {
+router.put("/:id", isLoggedIn, isOwner, validateListing, wrapAsync(async (req, res) => {
     // if (!req.body.listing) {
     //     throw new ExpressError(400, "Send valid data for Listing");
     // }
     let { id } = req.params;
+    // // Authorization : Already the Edit and Delete buttons are hidden for unauthourized accounts, but here, making the routes secure from server side from being able to access by api's.
+    // let listing = await Listing.findById(id);
+    // if (!listing.owner._id.equals(res.locals.currUser._id)) {
+    //     req.flash("error", `You do not have the permission to edit the listing ${listing.title}.`);
+    //     return res.redirect(`/listings/${id}`);
+    // }
+    // // now this authorization logic is moved to a middleware for better code management and reusability, and isOwner middleware is created in middleware.js file for that, which is then used here and in the delete route as well.
     await Listing.findByIdAndUpdate(id, { ...req.body.listing });
     req.flash("success", "Listing Updated!");
     res.redirect(`/listings/${id}`);                  // redirecting to show route instead of index
 }));
 
 //Delete Route
-router.delete("/:id", isLoggedIn, wrapAsync(async (req, res) => {
+router.delete("/:id", isLoggedIn, isOwner, wrapAsync(async (req, res) => {
     let { id } = req.params;
     await Listing.findByIdAndDelete(id);
     req.flash("success", "Listing Deleted!");
