+++ routes/review.js
@@ -6,20 +6,21 @@ const router = express.Router({ mergeParams: true });       // mergeParams: true
 const Listing = require("../models/listing.js");
 const Review = require("../models/review.js");
 const wrapAsync = require("../utils/wrapAsync.js");
-const ExpressError = require("../utils/ExpressError.js");
-const { reviewSchema } = require("../schema.js");
+// const ExpressError = require("../utils/ExpressError.js");
+// const { reviewSchema } = require("../schema.js");           // moved with middleware func
+const { validateReview } = require("../middleware.js");
 
 
-//middleware for validating review data using Joi schema
-const validateReview = (req, res, next) => {
-    let { error } = reviewSchema.validate(req.body);
-    if (error) {
-        let errMsg = error.details.map((el) => el.message).join(",");
-        throw new ExpressError(400, errMsg);
-    } else {
-        next();
-    }
-}
+// //middleware for validating review data using Joi schema
+// const validateReview = (req, res, next) => {
+//     let { error } = reviewSchema.validate(req.body);
+//     if (error) {
+//         let errMsg = error.details.map((el) => el.message).join(",");
+//         throw new ExpressError(400, errMsg);
+//     } else {
+//         next();
+//     }
+// }                             // moved to middleware file
 
 
 //Post Review Route
