+++ views/listings/show.ejs
@@ -8,22 +8,24 @@
             <div class="card col-6 offset-3 show-card listing-card">
                 <img src="<%= listing.image.url %>" class="card-img-top show-img" alt="listing_image">
                 <div class="card-body">
-                    <p class="card-text">
-                        <%= listing.description %> <br />
-                        &#8377;<%= listing.price.toLocaleString("en-IN") %> <br />
-                        <%= listing.location %> <br />
-                        <%= listing.country %> <br />
-                    </p>
+                    <p class="card-text">Owned by <i><%= listing.owner.username %></i></p>
+                    <p class="card-text"><%= listing.description %></p>
+                    <p class="card-text">&#8377;<%= listing.price.toLocaleString("en-IN") %></p>
+                    <p class="card-text"><%= listing.location %></p>
+                    <p class="card-text"><%= listing.country %></p>
+                    <!-- <br /> -->
                 </div>
             </div>
     </div>
-    <div class="btns">
-        <a href="/listings/<%=listing._id%>/edit" class="btn btn-dark col-1 offset-3 edit-btn">Edit</a>
-            
-        <form method="POST" action="/listings/<%=listing._id%>?_method=DELETE">
-            <button class="btn btn-dark offset-5">Delete</button>
-        </form>
-    </div>
+    <!-- < if (currUser && currUser._id.equals(listing.owner._id)) { %> -->
+        <div class="btns">
+            <a href="/listings/<%=listing._id%>/edit" class="btn btn-dark col-1 offset-3 edit-btn">Edit</a>
+                
+            <form method="POST" action="/listings/<%=listing._id%>?_method=DELETE">
+                <button class="btn btn-dark offset-5">Delete</button>
+            </form>
+        </div>
+    <!-- < } %> -->
 
     <div class="col-8 offset-3 mb-3">
         <hr />
